Refactor: Rename `vizard` module to `viz` and update import in `__init__.py`
