+++ src/chonkie/utils/__init__.py
@@ -1,6 +1,6 @@
 """Module for utility functions."""
 
 from .hub import Hubbie
-from .vizard import Visualizer
+from .viz import Visualizer
 
 __all__ = ["Hubbie", "Visualizer"]
